+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* feat: add static code diagnostics `avoid-global-state`.
 * chore: migrate from deprecated api
 * fix: fixed issue with type check in prefer-match-file-name
 * doc: add flutter favorite badge